+++ continual/conv.py
@@ -227,7 +227,11 @@ class _ConvCoNd(CoModule, _ConvNd):
         return output
 
     @property
-    def delay(self):
+    def receptive_field(self) -> int:
+        return self.kernel_size[0] + (self.kernel_size[0] - 1) * (self.dilation[0] - 1)
+
+    @property
+    def delay(self) -> int:
         return self.kernel_size[0] - 1 - self.padding[0]