+++ continual/module.py
@@ -89,6 +89,12 @@ class CoModule(ABC):
     Deriving from this class enforces that neccessary class methods are implemented
     """
 
+    receptive_field: int = 1
+    delay: int = 0
+    stride: int = 1
+    padding: int = 0
+    make_padding = torch.zeros_like
+
     def __init_subclass__(cls) -> None:
         CoModule._validate_class(cls)
 
@@ -112,10 +118,18 @@ class CoModule(ABC):
                 getattr(cls, fn, None)
             ), f"{cls} should implement a `{fn}` function which performs {description} to satisfy the CoModule interface."
 
-        assert hasattr(cls, "delay") and type(cls.delay) in {
-            int,
-            property,
-        }, f"{cls} should implement a `delay` property to satisfy the CoModule interface."
+        for prop in {"delay", "receptive_field"}:
+            assert type(getattr(cls, prop, None)) in {
+                int,
+                property,
+            }, f"{cls} should implement a `{prop}` property to satisfy the CoModule interface."
+
+        for prop in {"stride", "padding"}:
+            assert type(getattr(cls, prop, None)) in {
+                int,
+                property,
+                tuple,
+            }, f"{cls} should implement a `{prop}` property to satisfy the CoModule interface."
 
     @staticmethod
     def is_valid(module):
@@ -144,8 +158,6 @@ class CoModule(ABC):
     def clean_state(self):
         """Clean model state"""
         ...  # pragma: no cover
-
-    make_padding = torch.zeros_like
 
     def forward_step(
         self, input: Tensor, update_state=True